# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AR-VisuMathia game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlayerProgress {
  id: number;
  totalStars: number;
  totalCrystals: number;
  totalCoins: number;
  currentScene: number;
  scenesCompleted: number;
  totalPlayTimeSeconds: number;
  createdAt: string;
  updatedAt: string;
}

export interface Scene {
  id: number;
  sceneNumber: number;
  name: string;
  description: string;
  isUnlocked: boolean;
  isCompleted: boolean;
  starsEarned: number;
  maxStars: number;
  bestScore: number;
  playCount: number;
  category: string;
}

export interface SceneCompletionInput {
  starsEarned: number;
  score: number;
  playTimeSeconds: number;
}

export interface RewardSummary {
  totalStars: number;
  totalCrystals: number;
  totalCoins: number;
  badges: string[];
}

export interface RewardInput {
  stars: number;
  crystals: number;
  coins: number;
  /** @nullable */
  badge?: string | null;
}

export interface GameStats {
  totalScenesCompleted: number;
  totalStars: number;
  totalPlayTimeSeconds: number;
  averageStarsPerScene: number;
  mostRecentScene: number;
  completionPercentage: number;
}