# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AR-VisuMathia game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scenes/{sceneId}:
    get:
      operationId: getScene
      tags: [scenes]
      summary: Get scene details
      parameters:
        - name: sceneId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scene details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
        "404":
          description: Scene not found
  /scenes/{sceneId}/complete:
    post:
      operationId: completeScene
      tags: [scenes]
      summary: Mark a scene as completed
      parameters:
        - name: sceneId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SceneCompletionInput"
      responses:
        "200":
          description: Scene completion recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
...
components:
  schemas:
...
    Scene:
      type: object
      properties:
        id:
          type: integer
        sceneNumber:
          type: integer
        name:
          type: string
        description:
          type: string
        isUnlocked:
          type: boolean
        isCompleted:
          type: boolean
        starsEarned:
          type: integer
        maxStars:
          type: integer
        bestScore:
          type: integer
        playCount:
          type: integer
        category:
          type: string
      required:
        - id
        - sceneNumber
        - name
        - description
        - isUnlocked
        - isCompleted
        - starsEarned
        - maxStars
        - bestScore
        - playCount
        - category
    SceneCompletionInput:
      type: object
      properties:
        starsEarned:
          type: integer
        score:
          type: integer
        playTimeSeconds:
          type: integer
      required:
        - starsEarned
        - score
        - playTimeSeconds
...