# lib/api-zod/src/generated/types/sceneCompletionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AR-VisuMathia game API
 * OpenAPI spec version: 0.1.0
 */

export interface SceneCompletionInput {
  starsEarned: number;
  score: number;
  playTimeSeconds: number;
}